refactor: update loading-sentences component to import sentences correctly, adjust character-creator styling, and enhance login/signup forms with 8bit UI components
